fix struct
